pyproject.toml
[project]
name = "api-auto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.0",
    "faker>=37.5.3",
    "httpx>=0.28.1",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
]
